+++ Directory.Build.props
@@ -1,13 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project>
-	<!-- These properties will be shared for all projects -->
-	<PropertyGroup>
-		<VersionPrefix>4.400.0</VersionPrefix>
-		<VersionSuffix>
-		</VersionSuffix>
-		<VersionSuffix Condition=" '$(VersionSuffix)' != '' AND '$(BuildNumber)' != '' ">$(VersionSuffix)-$(BuildNumber)</VersionSuffix>
-	</PropertyGroup>
-	<PropertyGroup>
-		<TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-	</PropertyGroup>
+  <!-- These properties will be shared for all projects -->
+  <PropertyGroup>
+    <VersionPrefix>4.400.0</VersionPrefix>
+    <VersionSuffix></VersionSuffix>
+    <VersionSuffix Condition=" '$(VersionSuffix)' != '' AND '$(BuildNumber)' != '' ">$(VersionSuffix)-$(BuildNumber)</VersionSuffix>
+  </PropertyGroup>
+  <PropertyGroup>
+    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+  </PropertyGroup>
 </Project>
