+++ src/VirtoCommerce.GoogleEcommerceAnalyticsModule.Core/Models/GoogleAnalyticsSettings.cs
@@ -1,0 +1,8 @@
+namespace VirtoCommerce.GoogleEcommerceAnalyticsModule.Core.Models
+{
+    public class GoogleAnalyticsSettings
+    {
+        public bool EnableTracking { get; set; }
+        public string MeasurementId { get; set; }
+    }
+}
