+++ src/VirtoCommerce.GoogleEcommerceAnalyticsModule.Core/Services/IGoogleAnalyticsSettingsManager.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+using VirtoCommerce.GoogleEcommerceAnalyticsModule.Core.Models;
+
+namespace VirtoCommerce.GoogleEcommerceAnalyticsModule.Core.Services
+{
+    public interface IGoogleAnalyticsSettingsManager
+    {
+        Task<GoogleAnalyticsSettings> GetAsync(string storeId);
+    }
+}
