+++ src/VirtoCommerce.GoogleEcommerceAnalyticsModule.Core/VirtoCommerce.GoogleEcommerceAnalyticsModule.Core.csproj
@@ -7,15 +7,12 @@
     <IncludeSymbols>true</IncludeSymbols>
     <SymbolPackageFormat>snupkg</SymbolPackageFormat>
   </PropertyGroup>
-
   <PropertyGroup>
     <!-- Project is not a test project -->
     <SonarQubeTestProject>false</SonarQubeTestProject>
   </PropertyGroup>
-
   <ItemGroup>
     <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="All" />
-    <PackageReference Include="VirtoCommerce.Platform.Core" Version="3.200.0" />
+    <PackageReference Include="VirtoCommerce.Platform.Core" Version="3.400.0" />
   </ItemGroup>
-
 </Project>
