+++ src/VirtoCommerce.GoogleEcommerceAnalyticsModule.Web/module.manifest
@@ -1,12 +1,14 @@
-﻿﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <module xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
   <id>VirtoCommerce.GoogleEcommerceAnalytics</id>
   <version>4.400.0</version>
   <version-tag />
-  <platformVersion>3.200.0</platformVersion>
+
+  <platformVersion>3.400.0</platformVersion>
   <dependencies>
-    <dependency id="VirtoCommerce.Store" version="3.200.0" />
+    <dependency id="VirtoCommerce.Store" version="3.400.0" />
   </dependencies>
+
   <apps>
     <app id="ga4">
       <title>Google Analytics</title>
@@ -27,7 +29,7 @@
   <iconUrl>Modules/$(VirtoCommerce.GoogleEcommerceAnalytics)/Content/logo.png</iconUrl>
   <requireLicenseAcceptance>false</requireLicenseAcceptance>
   <releaseNotes>First version.</releaseNotes>
-  <copyright>Copyright © 2011-2022 Virto Commerce. All rights reserved</copyright>
+  <copyright>Copyright © 2011-2023 Virto Commerce. All rights reserved</copyright>
   <tags>analytics</tags>
   <assemblyFile>VirtoCommerce.GoogleEcommerceAnalyticsModule.Web.dll</assemblyFile>
   <moduleType>VirtoCommerce.GoogleEcommerceAnalyticsModule.Web.Module, VirtoCommerce.GoogleEcommerceAnalyticsModule.Web</moduleType>
